Feat(AuthModule): confirm password and password match validation decorator
Description:
Add a new decorator to verify if the given password and confirm password
are match. This decorator was developed in order to be used in the
register process but the register process development was canceled, so
the decorator remained as an Orphan decorator. Maybe it will be usefull
in the future updates.

Logs:
[Added][AuthModule] Add a new decorator to validate confirm password.