Fix(AuthModule): Fix persian and arabic numbers
Description:
There was a bug that when client entered numbers as persian or arrabic
system was not abale to handle them in the future request due to not
being abale to find match between english and persian or arabic numbers.
In order to fix this problem a new utility had been added to the system
to convert persian and arabic numbers to english.

Logs:
[Added][Utilities] Add number converter utilities.
[Modified][AuthModule] Use number utilities to fix phone numbers.