Refactor(Fix number utility): Fix typo in fix number utility file name
Description:
There was a typo in fix numbers utility file which was causing the file
to be unlike other files, so in this commit the file name has been
updated, Additionally there was an update in messages and the account
not found message has been moved from auth messages to not founded
messages.

Logs:
[Refactored][numberUtility] Update utility file name.
[Modified][Locales] Move account not found message from auth to not
founded messages.
[Modified][AuthModule] Update number utility file name and not found
messages.