Fix(General): Fix pagination not working properly
Description:
There was a bug in pagination process which was causing pagination to
not working properly and return data even if the current page was after
the maximum available total pages, this problem was due to pagination
data type being string instead of number. in order to fix this problem
the forgotthen plain to class transformer method from class-transformer
package has been added to the controller of the problematic APIs.